third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-with-host-pseudo-class.html

<!doctype html>
<html>
<head>
<script src="resources/shadow-dom.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body class="bodyClass">

  <template id="tmpl">
    <style>
      :host > .blue {
        color: blue;
      }
      :host > ::content > * {
        color: red;
      }
    </style>
    <div class="blue">I am blue!</div>
    <content></content>
  </template>

  <div id="host" class="hostClass">
    <div class="content">
      content
    </div>
  </div>

  <pre id="console"></pre>
</body>
<script>
description("Test for crbug.com/274063, cannot style ::content with a rule that includes :host.");

var host = document.querySelector('#host');
var root = host.createShadowRoot();
var template = document.querySelector('#tmpl');
root.appendChild(template.content);

shouldBe('window.getComputedStyle(document.querySelector("#host > .content")).color', '"rgb(255, 0, 0)"');
</script>
</html>
